fixed create config data
