+++ src/app/services/index.ts
@@ -24,3 +24,4 @@ export * from './plugin.service';
 export * from './packages-log.service';
 export * from './shared.service';
 export * from './roles.service';
+export * from './configuration-control.service';
